// front-end/src/types/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
      emails: {
        Row: {
          created_at: string | null;
          email: string;
        };
        Insert: {
          created_at?: string | null;
          email: string;
        };
        Update: {
          created_at?: string | null;
          email?: string;
        };
      };
      profiles: {
        Row: {
          active_index: string;
          created_at: string | null;
          id: string;
          max_pages: number;
          name: string | null;
          pages_uploaded: number;
          payment_type: string;
          picture_url: string | null;
          customer_id: string | null;
          index_status : string;
        };
        Insert: {
          active_index?: string;
          created_at?: string | null;
          id: string;
          max_pages: number;
          name?: string | null;
          pages_uploaded?: number;
          payment_type?: string;
          picture_url?: string | null;
        };
        Update: {
          active_index?: string;
          created_at?: string | null;
          id?: string;
          max_pages?: number;
          name?: string | null;
          pages_uploaded?: number;
          payment_type?: string;
          picture_url?: string | null;
        };
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
// ...
}